pyproject.toml
[project]
name = "scheduler-service"
version = "0.1.0"
description = ""
authors = [
    {name = "CoCongV", email = "cong.lv.yx@gmail.com"},
]
dependencies = [
    "fastapi~=0.95.0",
    "uvicorn~=0.22.0",
    "aioredis~=1.3",
    "click~=8.0",
    "asyncpg~=0.23",
    "tortoise-orm~=0.20.0",
    "passlib~=1.7",
    "pyjwt~=2.1",
    "pydantic~=1.10.0",
    "python-multipart~=0.0.6",
    "dramatiq~=1.14",
    "pika~=1.3.1",
    "motor~=2.4",
    "aiohttp~=3.7",
    "prompt-toolkit~=3.0",
    "cchardet~=2.1",
    "aiodns~=3.0",
    "brotlipy~=0.7",

]
requires-python = ">=3.13,<4.0"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/CoCongV/scheduler-service#"


[project.optional-dependencies]
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pylint~=2.8",
    "rope~=0.19",
    "requests~=2.25",
    "pytest-cov~=2.12",
    "pytest-sugar~=0.9",
    "pytest-mock~=3.6",
    "pytest~=6.2",
    "yapf~=0.31",
    "pytest-sanic~=1.7",
    "uvicorn~=0.14",
    "flake8~=3.9",
]